feat: add lazy app loading and window manager tests
