+++ .gitignore
@@ -1,0 +1,3 @@
+node_modules
+
+# test coverage etc could be added but at minimum
